refact: minor updates
