+++ web/beacon-app/src/features/topics/routes/TopicDetailPage.tsx
@@ -1,11 +1,9 @@
-import { t } from '@lingui/macro';
 import { Heading } from '@rotational/beacon-core';
 import invariant from 'invariant';
 import { useParams } from 'react-router-dom';
 
 import AppLayout from '@/components/layout/AppLayout';
 import DetailTooltip from '@/components/ui/Tooltip/DetailTooltip';
-import { formatDate } from '@/utils/formatDate';
 
 import AdvancedTopicPolicy from '../components/AdvancedTopicPolicy';
 import TopicQuery from '../components/TopicQuery';
@@ -13,29 +11,11 @@ import TopicQuickView from '../components/TopicQuickView';
 import TopicsBreadcrumbs from '../components/TopicsBreadcrumbs';
 import TopicSettings from '../components/TopicSettings';
 import { useFetchTopic } from '../hooks/useFetchTopic';
-
+import { getFormattedTopicData } from '../utils';
 const TopicDetailPage = () => {
   const param = useParams();
   const { id: topicID } = param;
   const { topic } = useFetchTopic(topicID as string);
-  const topicData = [
-    {
-      label: t`Topic ID`,
-      value: topic?.id,
-    },
-    {
-      label: t`Status`,
-      value: topic?.status,
-    },
-    {
-      label: t`Created`,
-      value: formatDate(new Date(topic?.created as string)),
-    },
-    {
-      label: t`Modified`,
-      value: formatDate(new Date(topic?.modified as string)),
-    },
-  ];
 
   invariant(topicID, 'topic id is required');
   return (
@@ -45,7 +25,7 @@ const TopicDetailPage = () => {
           <span className="mr-2" data-cy="topic-name">
             {topic?.topic_name}
           </span>
-          <DetailTooltip data={topicData} />
+          <DetailTooltip data={getFormattedTopicData(topic)} />
         </Heading>
         <TopicSettings />
       </div>
