--- web/beacon-app/src/features/topics/utils.tsx
@@ -1,29 +1,0 @@
-import { t } from '@lingui/macro';
-
-export const getDefaultTopicStats = () => {
-  return [
-    {
-      name: t`Online Publishers`,
-      value: 0,
-      units: '',
-    },
-    {
-      name: t`Online Subscribers`,
-      value: 0,
-      units: '',
-    },
-    {
-      name: t`Events`,
-      value: 0,
-    },
-    {
-      name: t`Data Storage`,
-      value: 0,
-      units: 'GB',
-    },
-  ];
-};
-
-export const getTopicStatsHeaders = () => {
-  return [t`Online Publishers`, t`Online Subscribers`, t`Events`, t`Data Storage`];
-};
